Move scenario folder into correct position
