+++ src/commands/commands.ts
@@ -4,7 +4,7 @@ import {Command} from "../types/command";
 import atis from "./basic/atis";
 import event from "./events/event";
 import license from "./basic/license";
-import scenario from "./events/scenario/scenario";
+import scenario from "./scenario/scenario";
 
 export default [
     help,
