+++ src/commands/scenario/common_options.ts
